Remove a test for a condition no longer permitted by dpkg (absolute link target).
The latest release of Dpkg will build the test package but cannot
unpack it.  The error message is:

    dpkg-source: error: pathname '[...]/debian/tests/asym2' points outside source root

The absolute link target, which is caught here, is exactly the
condition being tested. It is also an error tag in Lintian, and a
serious condition that should probably be flagged by Dpkg.

It's a little odd that Dpkg packaged the test in the first place.
Maybe it could error out when building something like it in the
future.

There is apparently another test using a homegrown way to assemble
packages that looks for this condition. That is why the tag is not
being removed now, but it may be in the future.

Gbp-Dch: ignore