Remove a test for a condition no longer permitted by dpkg (ignore link pointing outside source).
The latest release of Dpkg will build the test package but cannot
unpack it.  The error message is:

    dpkg-source: error: pathname '[...]/debian/source' points outside source root

The relative link target outside the package root, which is caught
here, is exactly the condition being tested. This test ensures that
the link is not followed. It is a serious condition that should
probably be flagged by Dpkg.

It's a little odd that Dpkg packaged the test in the first place.
Maybe it could error out when building something like it in the
future.

There is no tag to remove.

Gbp-Dch: ignore