Remove absolute link targets from tests that did not need them.
The latest release of Dpkg refuses to unpack some source packages. The
condition here are absolute link targets, which should not be
permitted is source packages. Lintian flagged them also as errors.

The Dpkg error messages were:

    dpkg-source: error: pathname '[...]/debian/tests/asym2' points outside source root

All three tests sprung from the same test and showed the same error.
Fortunately, the absolute link target was not related to the part
being tested.

It's a little odd that Dpkg packaged the tests in the first place.
Maybe it could error out when building something similar in the
future.

Gbp-Dch: ignore