Hold off on exempting JARs from breakout-link tag. (See: #963939)
It was brought to our attention that JAR files are often platform
independent and usually reside in /usr/share. That is also the
destination of the links being exempted.

There is an open question as to why the links at issue (from the
package eclipse-debian-helper) are located in /usr/lib.

This effectively reverts commit 6a6112f6 until the question is
resolved.