Fix tag-related URLs for new standalone HTML output format.
People often use the hyperlinks for tags to read explanations and
research potential remedies.

The new URLs include both the address lintian.d.o and the Lintian
version number, such as:

    https://lintian.debian.org/${lintan_version}/tags/${tagname}.html

The old URLs were relative to the document root, which did not include
any part of the fully qualified hostname.

Gbp-Dch: ignore