Switch to XS implementation of Text::Levenshtein for speed.
The C algorithm is lot faster. The speed up could be a factor of a
hundred or more.

The similar Text::Levenshtein::XS may be even faster. It offers a
maximum distance input that aborts the algorithm early when the
condition is met. Unfortunately, it was not available in Debian.

A better algorithm may be Text::Levenshtein::Damerau but Debian had
only the pure Perl implementation, and not XS. Damerau is superior
because transposition are counted as a single edit.

Gbp-Dch: ignore