Extend spell check to all parts of debhelper rule targets. (Closes: #963765)
The previous algorithm only checked the main dh command.  It did not
consider prefixes like 'execute_before' or 'execute_after', which are
just as likely to be misspelled.

The bug reporter in the present case wished to catch misspellings of
the 'override' prefix. This extends the Levenshtein analysis to the
full rule targets.

The algorithm was also improved that it now returns the closest match
rather than the first. It mattered in situations like these:

  -override_dh_installdebs -> override_dh_installdeb
  +override_dh_installdebs -> override_dh_installdocs

Includes a test case for the example from the inkscape package
mentioned in the bug, namely the mispelling of 'oeverride'.