Allow substring comparison when searching for ELF strings triggering operations.
Not all substrings will matched when anchoring to the beginning.  The
SHF_MERGE option in ld means that substrings are often created by
using a different index into the same static memory location.

This removes the absolute anchor at the beginning.

The regular expression provided by the caller should be modified to
include word or non-word boundaries.  In the case of file paths, which
start with a slash but end in a word character, \B was used in the
beginning and \b at the end.

The substring matching introduced with this commit is likely superior
when looking for substrings in a string spanning multiple commands.

Gbp-Dch: ignore