Add reminder to send forwardable DEP-3 patches upstream. (Closes: #755153)
Right now, this was only implemented for DEP-3 patches.

Not sure if Bug-Vendor also means that forwarding is not needed. The
DEP-3 specification [1] only stated:

    If the field is missing, its implicit value is "yes" if the
    "Bug" field is present, otherwise it's "no". The field is
    really required only if the patch is vendor specific, in that
    case its value should be "not-needed" to indicate that the
    patch must not be forwarded upstream (whereas "no" simply
    means that it has not yet been done).

[1] https://dep-team.pages.debian.net/deps/dep3/

Provides three test cases; two regular and one false positive.
Regenerates profiles.

A bunch of tests for output formats generate automatic Debian patches,
which are not marked as forwarded.  The extra tags appear here.