In one test, ignore tags seen only on some architectures; causes autopkgtest regression on arm64.
On arm64, this test fails with:

    Tags do not match

    --- ../../autopkgtest_tmp/testsuite/eval/checks/files/contents/bin-sbin-confusion-in-elf/tags.specified.calibrated
    +++ ../../autopkgtest_tmp/testsuite/eval/checks/files/contents/bin-sbin-confusion-in-elf/tags.actual.parsed
    +bin-sbin-confusion-in-elf (binary): bin-sbin-mismatch usr/bin/calls-sbin sbin/our-script -> usr/bin/our-script
    +bin-sbin-confusion-in-elf (binary): bin-sbin-mismatch usr/bin/calls-sbin bin/our-script -> usr/bin/our-script

    Failed test 'Lintian passes for bin-sbin-confusion-in-elf'

It is in fact the correct result, but on amd64 (and possibly other
architectures), ld is configured to reuse strings with SHF_MERGE as
noted in commit d446ff0d.

Removes the two correct tags to make the result uniformly wrong across
architectures.

Fixes an autopkgtest regression on arm64.