For testsuites, parse d/control instead *.dsc; drop tag inconsistent-testsuite-field.
The field Testsuite: autopkgtest is automatically added to the *.dsc
when d/tests/control is present. Likewise, a standard d/tests/control
is supplied when the field is present in d/control.

The tag inconsistent-testsuite-field used to check for that mechanism,
but Lintian is not the dpkg testsuite. Instead, the check now looks at
the maintainer-provided data in d/control.

The tag was renamed to missing-tests-control. It is now used when the
field Testsuite: autopkgtest is declared but no d/tests/control is
present (unless a specialized test suite like autopkgtest-pkg-perl
provides it).  Renames and repurposes the associated test case.

Removes the test case testsuite-autopkgtest-missing, which tests a
condition that can no longer occur. Also removes the tag it tests. It
can no longer occur, either.

Also updates one test case to use the new field name Testsuite instead
of the older XS-Testsuite.