Use new Deb822 facility for all primary package control files.
Each Lintian::Processable now offers the fields in its primary control
file via the ->fields method.

The method supplies a Lintian:Deb822::Section that stores and accesses
field values in a case-invariant way.

The code being replaced was already the basis for the new facilites in
their separate classes. This completes the transfer.

Gbp-Dch: ignore