Use new Deb822 facility for the fields in d/control.
Source processables offer special methods to access fields from
d/control. The code being replaced was (or should be) identical in
function to the new classes being used from now on.

Two tags concatenated package names as part of their context messages.
The messages were constant for purposes of the test suite but relied
on the order of appearance in a file, which was disturbed by the use
of hashes as interim data structures.

A sort was therefore imposed on the package names before output, yet
that order ended up being slightly different. The changes to expected
tags in one test reflect the new order.

Gbp-Dch: ignore