Add two classification tags for Salsa CI detection.
Lintian can now integrate with Salsa CI, but candidates for deployment
are hard to find. Until now, the state of the art was to examine:

    https://codesearch.debian.net/search?q=salsa-ci-team%2Fpipeline%2Fraw%2Fmaster%2Fsalsa-ci.yml&literal=1

Starting with this commit, Lintian will issue two classification tags:

One for the presence of a likely CI specification. The names for such
files can be configured per project and are not standardized. The
names of candidates were taken from this Salsa issue:

    https://salsa.debian.org/salsa-ci-team/pipeline/-/issues/86

The second tag shows included files. Such files often point to
standard setups provided by the Salsa CI team for the benefit of the
community. They generate Lintian output in a HTML format that can be
viewed directly in Salsa.

Both tags use the new (and optional) name-spaced format.

Provides a test case. Regenerates profiles.