Store tag aliases in Moo array references, not in arrays; fix broad failure of tag renaming mechanism.
This error caused almost the entire renaming mechanism to fail.

The bug was discovered when rename-tag did not show for unknown-field
in postgresql-13_13~beta2-2_amd64.deb. The circumstances are further
documented in the nearby commit that fixes it.

Thanks to Christoph Berg for bringing the issue to our attention!

Gbp-Dch: ignore