Prohibit commas as separator for old tag names in the Renamed-From field in tag definitions.
Caused parsing errors that led to a partial breakdown of the tag
renaming mechanism.

The bug was discovered when rename-tag did not show for unknown-field
in postgresql-13_13~beta2-2_amd64.deb. The circumstances are further
documented in the nearby commit that fixes it.

Thanks to Christoph Berg for reporting this on IRC!

Gbp-Dch: ignore