Copyright check: add another template.
dh-make-perl since version 0.106 adds a template for missing upstream
copyright years if it can't find them in the upstream metadata.

Check for this template as well and issue the helper-templates-in-copyright
tag if found.

/*
Code-wise this could also put a few lines up to the other template checks,
but it might be easier to adjust in the future if it's grouped with the
other check for a dh-make-perl-generated boilerplate.

Code in dh-make-perl:
lib/DhMakePerl/Command/Packaging.pm:        'Copyright: ' . ( $cprt_year ? $cprt_year : '<INSERT COPYRIGHT YEAR(S) HERE>' )
*/