In Salsa CI spec, locate includes at all levels of YAML file.
The original check assumed that include directives would be located as
an array at a particular level in the hierarchy, but that is not
necessarily so.

Spotted via an error in PostgreSQL, which places the directive as a
scalar at the root level.

    Can't use string ("https://salsa.debian.org/postgre"...)
    as an ARRAY ref while "strict refs" in use at
    /usr/share/lintian/checks/continuous-integration/salsa.pm
    line 65.

Provides a test case for scalar at the document root.

Gbp-Dch: ignore