Move Salsa CI specification to d/salsa-ci.yml.
About a year ago, the Salsa CI team changed the recommended name for
CI specifications to d/salsa-ci.yml. It created some controversy [1]
that was ultimately settled in the team's favor.

[1] https://salsa.debian.org/salsa-ci-team/pipeline/-/issues/86

Like the team that filed in the Salsa issue, Lintian was an early
adopter. We used an even older name.

A new check called continiuous-integration/salsa looks for four file
names as candidates for Salsa CI and will soon provide statistics.
That may help to analyze the situation further.