Add test case for email address with missing closing bracket. (See: #965335)
The module Email::Address::XS does not complain about missing closing
brackets after an email address as long as the contact information can
be parsed correctly. It is unclear if that is a bug.

This adds a test case for that situation. Thanks to Mechthilde
Stehmann for bringing the issue to our attention!