Replace Lintian::Util::do_fork with the equivalent IO::Async code.
Uses stock library routines instead of the custom procedures offered
in Lintian::Util.

This commit is also an attempt to track down a Heisenbug that showed
up when the modules Data::DPath and Scalar::Util were added to the
check continuous-integration/salsa.

That uncovered what is likely a race condition elsewhere in Lintian.

One pointer in the Heisenbug was that the standard SIGCHILD handler
had been removed without a replacement. It may or may not have
happened in Lintian::Util::do_fork.

This commit eliminates one possibility. The switch in this check to
IO::Async was scheduled anyway.

Gbp-Dch: ignore