Run each check in a separate process; execute in alphabetic order.
This fixes a Heisenbug that appeared after Data::DPath was used in the
check continuous-integration/salsa.

The problem was probably caused by the mixing of IO::Async routines
and other calls such as system() in various checks. The call to
safe_qx in the check fields/multi-arch seemed to fail because no
SIGCHLD handler was installed.

The isolation slows the execution a bit.  Lintian generates a lot of
classification tags, which take time to serialize.  (The Processable
data structure is closed upon, so no part of it is serialized.)

An big advantage, however, should be memory usage going forward.  The
improved isolation between checks should free up memory when each
check is done.

The last point is hard to quantify without further research but the
checks slurp a lot of files, especially because often UTF-8 handling
requires it.

Also executes the checks in alphatic order. The order was previously
unsorted.

Checks do not run in parallel. An Ia::Function with its automated
worker processes would be most convenient, but the last time it was
attempted an extra Ia::Channel was needed, and Ia::Function did not
offer it.

Gbp-Dch: ignore