Account properly for license identifiers in d/copyright; split wildcard matching from license parsing. (Closes: #779676)
Based on the number of open bugs, the copyright check was in urgent
need of refurbishment. This commit separates license processing from
wildcard matching.

The check will eventually be split into multiple sub-checks.

This commit also made it possible to close an old bug. The committer
is not sure it is a good idea to allow the reuse of identifiers for
different licenses in the same d/copyright file, but the opinion among
other commenters was heavily in favor.

Includes the false positive test case from the bug in a slighly
modified form (but with annotations to explain the differences).

Short license identifiers from the first lines of license paragraphs
are now accounted properly. Later appearances no longer overwrite
earlier appearances. Arrays are kept instead.

The extra tags appearing in the test suite are true positives.  They
were previously not identified but are proper.