Disregard watch files containing only comments. (Closes: #965385)
Makes the exact edit suggested by the bug filer. Thanks to gregor
hermann for bringing this issue to our attention!

Provides a test case for the false positive described in the bug.

Also removes four false positive tags from the test suite. The watch
files are either empty or blank.