Fix wildcard-matches-nothing-in-dep5-copyright. (Closes: #888001)
The tag may not have existed when the bug was filed in 2014, but it
was also not working correctly. This fixes the tag and the bug.

The bug calls for a tag named inexistent-file-in-dep5-copyright, which
it actually the same as wildcard-matches-nothing-in-dep5-copyright.
The latter is merely named better. The Files paragraphs list globbing
patterns.

Now the tag works as intended.

The bug reporter furthermore asked for the mismatch condition to be
split from unused-file-paragraph-in-dep5-copyright, but those two are
too different. Please examine the test suite if you disagree.

While the 'unused' tag shows when (and only when) all wildcards in a
paragraph suffer from mismatches, the unused condition also indicates
a potential license confusion.

Of course, the bug reporter is also correct in that both tags occur
together when Files paragraphs with single wildcards (which are
common) produce no matches. Let's see how this fix performs in the
wild.

Perhaps unused-file-paragraph-in-dep5-copyright should be silenced
when a wildcard-matches-nothing-in-dep5-copyright.

Includes the test case from the bug report with a wildcard for a file
that does not exist:

    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=888001#33

As expected, the test triggers both tags.

This commit also improves the line numbers in tag messages. Now the
context points to the Files field rather than the beginning of the
Deb822 section (aka the 'paragraph').