Also consider pseudo-headers allowed by DEP-3 when examining patches. (Closes: #966024)
As the bug reporter speculated already, the applicable check did
indeed not parse the pseudo-headers envisioned by DEP-3. This commit
implements it.

We keep the last mention of each field whenever they are mentioned
multiple times.

The package i3-gaps mentioned in the bug report could not be located,
but this commit provides a false-positive test case.

Thanks to Raphaël Hertzog for bringing the issue to our attention.