Analyze matching behavior and relative order of globbing patterns in d/copyright. (Closes: #905747)
This commit introduces three new tags. Two of them point out the
issues mentioned in the bug. The third addresses are related condition
for using the same wildcard in multiple paragraphs.

The tag globbing-patterns-out-of-order identifies issue (1) from the
bug. It indicates that a pattern specified later took precendence, by
virtue of the way DEP-5 copyright files are organized, over an earlier
pattern for a particular file but had a lower directory depth.

From a user's perspective, it means to reorder the paragraphs. The tag
should probably be an error instead of a warning, but it is new.

Meanwhile, the tag redundant-globbing-patterns identifies issue (2)
from the bug in a general manner. It indicates that two patterns in
the same paragraph matched the same file. That is redundant because
the same license applies to both patterns.

The third tag duplicate-globbing-pattern pops up when the same pattern
was used in multiple but different paragraphs. It indicates a possible
confusion about which license should apply.

Provides three test cases; one for each tag. Two test cases were
derived directly from the bug (with small modifications as explained
in comment):

    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=905747#5

The logic of the check has been much improved, and many small fixes
were made.  Unfortunately, the data structures are not as elegant as
they should be. Efforts to make their handling less convoluted are
under way.

Also, the check is currently not very good at preventing trickle down
tags.

For example, wildcard-matches-nothing-in-dep5-copyright can trigger
unused-file-paragraph-in-dep5-copyright when there is only one
wildcard. It was discussed briefly in Bug#888001 and commit 38c6284f.
The behavior is visible in the test wildcards-out-of-order in the test
suite. It is a bit cumbersome to the eye.

The trickle down effects will be addressed gradually as the data
structures are simplified.