Convert unused-override to a more intuitive algorithm.
No issues were reported but the code found unused overrides by
traversing those that had been used. The mechanism was hard to
understand.

The new algorithm instead looks at declared overrides and decides
which ones were used. Then it locates those that were not.

In the opinion of the author, the new version reads better. It also
prepares the code for the upcoming mismatched-override.

Gbp-Dch: ignore