Flag overrides with mismatched context separately from those not used.
The tag unused-override sometimes caused confusion when a context
changed. The 'unused' wording suggested the underlying condition had
been fixed. Instead the context had changed, and the override had to
be adjusted.

A new tag mismatched-override is now shown when an override could not
be used while the targeted tag was still issued. Clearly, the context
did not match.

The new tag is a warning. That is a level above unusued-override.  It
was part of the original suggestion.

While the level may seem alarmist, a maintainer would be even more
likely to see mismatched-override before the targeted tag if it were
an error.  Let's see how that performs in the wild.

The new tag cannot be overridden.

At the time of writing, the new tag was seen in the following packages
(and probably others):

    psychopy_2020.1.3+dfsg-2_all.deb
    simplesamlphp_1.18.7-1_all.deb
    sogo_4.3.0-1.dsc
    swe-standard-data_00004-1_all.deb

Provides two new test cases; one for mismatched-override, and one for
unused-override. Solves one untested tag.

Thanks to Christoph Berg for the suggestion!