Set CCACHE_DISABLE in t/bin/build-test-packages; otherwise standard Salsa CI pipelines willfail.
This commit disables ccache when building test packages. It is a
precursor to using the standard pipelines published by the Salsa CI
team.  Lintian will soon use these standard pipelines

Due to a request from the MariaDB maintainer, the pipeline installs
ccache. For some projects, that leads to enormous speedups:

    https://salsa.debian.org/salsa-ci-team/pipeline/-/issues/27

For Lintian there is probably no benefit. There is only some C code in
test packages, and those artifacts are expected to be cached
separately. (It will not work immediately after the switch.)

Our script that builds the test packages tries to avoid any variation
and sanitizes the environment. That, however, makes the variable
CCACHE_DIR unavailable. Gcc then tries to use the default value
$HOME/.ccache mentioned in the documentation:

    https://ccache.dev/manual/3.7.11.html

In the standard Salsa CI pipelines that location is not writable. The
builds for test packages fail with:

    ccache: error: Failed to create directory /home/salsa-ci/.ccache/tmp: Permission denied

This only happens in test packages that use gcc. None of this occurred
in Lintian's custom CI setup, as ccache was not installed.

This commit sets the variable CCACHE_DISABLE when building test
packages in an otherwise sanitized environment. It disables ccache
when it is installed, and has no effect otherwise.

More information may be available here:

    https://salsa.debian.org/salsa-ci-team/pipeline/-/issues/164