Use the Salsa CI team's standard pipeline instead of Lintian's custom CI setup.
The main motivation for this switch was that the standard pipeline
builds Debian packages. The Salsa CI team then offers other extra
pipeline jobs that far exceeds the CI testing we do currently.

In particular, this commmit makes Lintian tags and piuparts results
available.

The artifact caching we implemented with great care still has to be
ported to the new setup. That will have to happen in d/rules, but it
will not work at the moment because the docker pipeline blocks all
environment variables that do not start with DEB*.

The variable we would use to detect the CI setup do not start with
that prefix:

    https://docs.gitlab.com/ee/ci/variables/predefined_variables.html

The Salsa CI team is working on having that restriction relaxed.

The new setup also does not run any perltidy and perlcritic tests.
They were always disabled in our standard package builds to prevent
unexpected style failures, which can happen when new perltidy or
perlcritic versions are released.

The style tests will also be enabled once we can detect the Salsa CI
environment in d/rules.

For now, please run the full test suite t/bin/runtest on your machine
locally before committing. It will run perltidy and perlcritic for
you.