In another round of tests, do not create files with different ownership or group unless needed.
All these tests fail in the second build under reprotest with this
or similar errors:

    chown: changing ownership of 'debian/game/usr/games/script': Invalid argument

Many of the ownership changes were not being used and only ended up
where they were because the tests were duplicated when checks were
split.

Removes two tests that had no diagnostic value after the ownership
changes were removed from them.  Drops ownership changes from all
tests named files-bad-perm-owner.

It made no sense to move the ownership changes to the maintainer
scripts. It would sidestep the issue with reprotest but not trigger
the tags, which is the purpose of the test packages.

Places the untested tags on the untested list, which moves them a step
closer to removal. Their functionality is probably better checked in
piuparts, which installs packages and also runs maintainer scripts.

Gbp-Dch: ignore