Wait for each lexgrog and man; convert documentation/manual check to IPC::Run3. (Closes: #966072)
No longer forks an uncontrollable number of processes. The use of IPC::Run3
further reduces the number of open file descriptors over IO::Async, although
that should no longer matter.

This fix is untested. Please reopen the bug if the issue persists.

Part of a series to convert all checks to IPC::Run3.