In check files/compressed/lz, do not stat(2) lzip command for every file in the index.
It is not necessary to look for the lzip command for each file in the index.
The stat call is particularly expensive. One time at the beginning is fine.

Moo provides a suitable mechanism via lazy loading.

Also, please note that all modules should be included above the 'use Moo'
line, per the author's recommendation. [1]

[1] https://metacpan.org/pod/Moo#CLEANING-UP-IMPORTS