Run no tests when building; reduce Build-Depends; drop nocheck build profile.
Testing Lintian is paramount, but doing so every time a package is built is
excessive. In the standard CI pipeline on Salsa, the ~1200 tests ran seven
times over, for every commit.

Aside from the waste of resources, time was also a consideration. The Salsa
CI pipeline took 275 minutes.

The most extreme case was the reprotest job, which usually passed narrowly
under the three-hour job limit, at 175 minutes, and occasionally timed out.
Now that job takes 2 minutes 50 seconds.

This commit vastly simplifies Lintian's build prerequisites. Many were
listed in Depends already. Those that were not are now part of the Depends
field for the autopkgtest suite.

Drops the nocheck build profile. It is no longer needed.

It would probably be worthwhile to add autopkgtest pipelines for testing and
buster-backports. On those releases, the tests do not run at the moment.