Announce LINTIAN version in d/rules via echo.
It is not possible to call ./frontend/lintian sucessfully from the source
tree after INIT_ROOT was hardcoded to the system path /usr/share/lintian/.

In the current Salsa CI setup, it provoked the error message:

    lintian is not available
     at /builds/lintian/lintian/debian/output/lintian-2.86.0+git+salsaci/debian/lintian/usr/share/lintian/frontend/lintian line 161.

On systems with an older versions of Lintian installed, the following
message is seen instead:

    /usr/share/lintian/commands/lintian.pm does not define a main sub
    at /lcl/lechner/lintian/git/debian/lintian/usr/share/lintian/frontend/lintian line 145.

Here, the code that is about to be packaged called the installed version of
Lintian. The file commands/lintian.pm was outdated, however, and did not
declare a 'package' as described in commit 4e80dbc4. The ->can mechanism
from UNIVERSAL therefore failed.

It is unclear why the call in d/rules worked until now, and for so long.

Announces the version via echo instead.

Gbp-Dch: ignore