In Deb822 facility, always return trimmed values.
The new Deb822 facility may trim field values less aggressively than the
previous implementation with plain Perl hashes. This returns trimmed field
values for the ->value method by default.

To remove internal newlines please use the ->unfolded_value method.

Furthermore provides a new ->untrimmed_value method. It is used for some
formatting tags, such as multiline-field, and is needed for fields in which
the first line has a different meaning, such as License or Description. That
line may not be empty.

The condition cannot be detected with a trimmed field value.

Gbp-Dch: ignore