Install library modules to /usr/share/lintian/lib.
Perl's FindBin can easily find the location, which is always a sibling of
the script folder ./bin. The goal is to ultimately stop shipping the modules
in the system path.

The custom location will prevent Lintian from loading conflicting modules
from the system path when in fact run from an unpacked source tree.

The only known consumer of our modules is libconfig-model-dpkg-perl. It
loads dates from data/standard-versions/release-dates via Lintian::Data
(which requires a Lintian profile).

That package should get the dates from the debian-policy changelog instead.
It can be found here:

    https://salsa.debian.org/dbnpolicy/policy/-/blob/master/debian/changelog