Move executables from ./frontend to ./bin; set indirect links from /usr/bin.
The current infrastructure to locate Lintian modules and support files is
unreliable and too complicated. Perl offers a simple way with FindBin. To
that end, all executables---including those for testing---will be moved.

Each program can then find modules and support files in exactly the same
way, whether Lintian is installed or run from a source tree.

The use of the ./bin folder is not required. The name is changed merely for
convention.

The three programs that are currently dispatched by the frontend will be
permanently separated. The indirect links from /usr/bin will continue to
work afterwards, as well.