Move test executables from t/bin to ./bin.
The current infrastructure to locate Lintian modules and support files is
unreliable and too complicated. Perl offers a simple way with FindBin.

With this commit, all executables were moved into the same folder. Each
program can now find modules and support files the same way, whether Lintian
is installed or run from a source tree.

The use of the ./bin folder is not required. The name is merely for
convention.