Use Perl's FindBin to reliably detect the base directory for Lintian.
The new environment variable LINTIAN_BASE combines the functionality of
LINTIAN_ROOT and LINTIAN_TEST_ROOT, which no longer exist.

The reason for two variables was to allow for a version difference, but
under autopkgtest that is not needed.

When LINTIAN_TEST_INSTALLED is set to 'yes', the test suite will now use the
installed version as much as possible---even when loading some modules.

The Lintian executable being tested is now in LINTIAN_UNDER_TEST.