Provide a script to download policy release dates. (See: #968000, #968011)
The package libconfig-model-dpkg-perl uses Lintian internal modules, which
will no longer be installed in the Perl system path. This script may offer
the maintainers another way to get the informaton they require.

The package loads a default Lintian profile and then uses Lintian::Data to
access data/standards-version/release-dates. This script downloads the
changelog from the policy editors and extracts the same information.

the script obtains the release dates from their origin. It is superior to
any use of Lintian for that purpose.

Gbp-Dch: ignore