Announce left-over processes before exiting.
Most of the time, left over processes are done already and merely have an
unreaped exit status (so-called zombies). In that case, the table is not
very interesting.

Compared to the previous code, this code also provides greater details for
processes that are still running, although that should never occur during
normal operation.

Gbp-Dch: ignore