Add unzip to Depends in d/control; drop from d/tests/control. (Closes: 968041)
The previous implementation of safe_qx did not always distinguish between a
non-zero exit code and when a program was not installed. The reporting party
probably did not have 'unzip' installed.

I could not remove that package locally to test that hypothesis. It was
required by too many packages. Please confirm with a brief bug amendment
that the error disappears when unzip is installed.

Please reopen if the bug was closed in error.

Thanks to Sudip Mukherjee for reporting this issue!