Do not show hyperlinks unless showing color is also appropriate. (See: 968067)
Something came up when working on the bug: Hyperlinks were shown even when
the output was piped. Now that only happens when --color=always.

The solution is not ideal because now hyperlinks only show with color, but
they are probably used with color most of the time, anyway.

Gbp-Dch: ignore