Further reduce reliance on IO::Async in Lintian proper and run-time utilities.
The Lintian executable has too many checks and draws on too many third-party
modules to allow the general use of IO::Async. The module works great in the
test suite but in Lintian proper it contributed to mysterious bugs.

The mystery bugs occurred most likely because some modules spawned processes
with methods other than IO::Async. That caused conflicts because IO::Async
replaces the SIGCHLD handler.

A decision was therefore made to reduce the reliance on IO::Async in Lintian
proper. In many cases, IPC::Run3 was sufficient.

The complex operations when indices are created from tar files  may have to
be implemented with more basic components such as IO::Handle and IPC::Open3.

Gbp-Dch: ignore