Drop .pl extension from generate-profiles; use advanced Deb822 facilities.
According to policy 10.4, the script's name should not include the .pl
suffix. It was removed here.

Additional build prerequisites are not required for the full Deb822
facilities. The script is mentioned in d/rules, but it does not run
automatically.

Otherwise, modules such as Moo and List::Compare would be required.

Gbp-Dch: ignore