Un-deprecate /etc/lintianrc; remove file on upgrade and ship in doc/examples. (Closes: #968326)
Due to a rewrite of how configuration files were handled, the file
/etc/lintianrc was no longer read. The manual page mentioned that the
location became obsolete in version 2.5.12, released in April of 2013.

After consultation, the location /etc/lintianrc was un-deprecated. The XDG
location in /etc/xdg is less well-known. It was also unclear to the author
how XDG_CONFIG_DIRS could be considered.

Stops shipping a sample config file in /etc and places the file in
/usr/share/docs/lintian/examples instead.

On upgrade, removes /etc/lintianrc if the file was unchanged and installed
by a version prior to 2.90.0. The use of rm_conffile was based on dbus.

In the manual page, removes the deprecation comments and suggests the
user-based XDG location. States that files are searched in the order listed.
Corrects the order with respect to XDG_DIR/lintian/lintianrc, which was
actually preferred over ~/.lintianrc.

Thanks to Simon McVittie, Bastian Blank and others for providing guidance
and specific pointers, especially to dpkg-maintscript-helper(1).

Thanks to Joao Eriberto Mota Filho for reporting this issue! Please reopen
the bug if it was closed in error.