For the first time, unpack tarballs in orig index; use centralized streaming unpacker.
The files in the orig index were previously not unpacked, but their presence
is expected to be beneficial to the cruft check, and perhaps others.

The cruft check disabled some tags in source format 3.0 (quilt) because no
consolidated diffstat information was available for all Debian changes
together. That information can, however, be made available by looking at the
file differences between the patched and the orig indices.

From now on, the contents of orig source tarballs are readily available for
a full examination in checks. This commit uses the centralized and trusted
streaming unpacker.

Due to the complexities of handling multiple source components, it is
possible that the new code does not work entirely as expected.

The code strips the common prefix, which standard in tarballs, only from
index names and not from unpacked file paths. As far as we know, however,
nothing in Lintian expects unpacked file paths to correspond to index names
in any particular fashion.

All file references in Lintian are resolved via index names, which resemble
the identifier in the tarball. The unpacked file path is used only to access
files when standard methods to examine them are insufficient.

Adds a subfolder facility to the Index base class for the proper separation
of orig source components when files are extracted.