Issue repeated-path-segments only directories only.
Until now, the tag was issued for files. That produced too many tags when
folders were well-populated. This commit consolidates those paths into one
tag for the offending folder.

It is a huge improvement for the tag, the utility of which is disputed. This
makes the tag much less of a nuisance and perhaps assuages those who find it
unnecessary.

In the package awesome-doc, for example, the repetition of the path segment
'doc' in usr/share/doc/awesome/doc/ triggered the tag a whopping 325 times.

Now Lintian produces the following output:

% bin/lintian /mirror/debian/pool/main/a/awesome/awesome-doc_4.3-5_all.deb
P: awesome-doc: repeated-path-segment doc usr/share/doc/awesome/doc/

Thanks to deki for pointing it out!