Provide a Perl-native implementation to index and extract tar files in parallel. (Closes: #968611)
Resolves the Heisenbug that has troubled Lintian for so long. IO::Async
replaces the SIGCHILD handler. It works great in the test suite but not in
Lintian. The reason is unknown.

A check probably forks children in a third-party module via an unrelated
mechanism, which is incompatible with IO::Async.

With this commit, Lintian no longer uses any routines from IO::Async.

The bug fix can be confirmed by running this command in a source directory
with built tests:

    bin/lintian debian/test-out/packages/checks/apache2/*/*.changes

The command no longer aborts but produces a lengthy list of hints instead.